add 3 seconds in eta time

+++ apps/read_email/tasks_expires.py
@@ -8,15 +8,18 @@ from .models import Order
 def deactivate_expired_order(order_id):
     logging.info(f'####### Время удаление {order_id} #######')
     try:
-        order = Order.objects.get(pk=order_id)
-        if order.user is None:
-            logging.info(f"Время действия заказа: {order.id} истекло в: {order.this_posting_expires_est}")
-            order.delete()
-            logging.info(f"Заказ с номером {order_id} удален")
+        order = Order.objects.filter(pk=order_id).first()
+        if order:
+            if order.user is None:
+                logging.info(f"Время действия заказа: {order.id} истекло в: {order.this_posting_expires_est}")
+                order.delete()
+                logging.info(f"Заказ с номером {order_id} удален")
+            else:
+                logging.info(f"Время действия заказа {order.order_number} истекло. Заказ перемещен в историю")
+                order.is_active = False
+                order.save()
         else:
-            logging.info(f"Время действия заказа {order.order_number} истекло. Заказ перемещен в историю")
-            order.is_active = False
-            order.save()
-    except Order.DoesNotExist as e:
-        logging.error(f'Ошибка на order: {order_id} deactivate_expired_order {e}')
-        raise ValueError(f"No found order {order_id} deactivate_expired_order {e}")
+            logging.info(f'Order уже удален')
+    except Exception as e:
+        logging.error(f'Error in order: deactivate_expired_order {str(e)}')
+        raise ValueError(f"Error in order deactivate_expired_order {str(e)}")
